feat(suite): store only active network fees
