+++ packages/suite/src/actions/suite/__tests__/initAction.test.ts
@@ -12,7 +12,7 @@ import {
     periodicCheckTokenDefinitionsThunk,
 } from '@suite-common/token-definitions';
 import {
-    blockchainActions,
+    feesActions,
     fetchFiatRatesThunk,
     initBlockchainThunk,
     initDevices,
@@ -99,7 +99,7 @@ const fixtures: Fixture[] = [
             connectInitThunk.fulfilled.type,
             initBlockchainThunk.pending.type,
             preloadFeeInfoThunk.pending.type,
-            blockchainActions.updateFee.type,
+            feesActions.updateFee.type,
             preloadFeeInfoThunk.fulfilled.type,
             messageSystemActions.fetchSuccessUpdate.type,
             fetchConfigThunk.fulfilled.type,
@@ -143,7 +143,7 @@ const fixtures: Fixture[] = [
             connectInitThunk.fulfilled.type,
             initBlockchainThunk.pending.type,
             preloadFeeInfoThunk.pending.type,
-            blockchainActions.updateFee.type,
+            feesActions.updateFee.type,
             preloadFeeInfoThunk.fulfilled.type,
             messageSystemActions.fetchSuccessUpdate.type,
             fetchConfigThunk.fulfilled.type,
@@ -188,7 +188,7 @@ const fixtures: Fixture[] = [
             connectInitThunk.fulfilled.type,
             initBlockchainThunk.pending.type,
             preloadFeeInfoThunk.pending.type,
-            blockchainActions.updateFee.type,
+            feesActions.updateFee.type,
             preloadFeeInfoThunk.fulfilled.type,
             messageSystemActions.fetchSuccessUpdate.type,
             fetchConfigThunk.fulfilled.type,
