+++ packages/suite/src/hooks/wallet/useRbfForm.ts
@@ -3,6 +3,7 @@ import { useForm } from 'react-hook-form';
 
 import type { NetworkType } from '@suite-common/wallet-config';
 import { DEFAULT_OPRETURN, DEFAULT_PAYMENT, DEFAULT_VALUES } from '@suite-common/wallet-constants';
+import { DEFAULT_FEE_INFO } from '@suite-common/wallet-core';
 import {
     ChainedTransactions,
     FeeInfo,
@@ -98,7 +99,9 @@ const useRbfState = ({ selectedAccount, rbfParams, chainedTxs }: UseRbfProps) =>
     const { shouldSendInSats } = useBitcoinAmountUnit(account.symbol);
 
     return useMemo(() => {
-        const feeInfo = getRbfFeeInfo(account.networkType, symbolFees, rbfParams);
+        const rbfFeeInfo = symbolFees
+            ? getRbfFeeInfo(account.networkType, symbolFees, rbfParams)
+            : DEFAULT_FEE_INFO;
         // filter out utxos generated by this transaction
         const otherUtxo = (account.utxo || []).filter(input => input.txid !== rbfParams.txid);
         // filter out utxos with anonymity level below target and currently registered
@@ -164,7 +167,7 @@ const useRbfState = ({ selectedAccount, rbfParams, chainedTxs }: UseRbfProps) =>
         return {
             account: rbfAccount,
             network,
-            feeInfo,
+            feeInfo: rbfFeeInfo,
             coinjoinRegisteredUtxos,
             chainedTxs,
             shouldSendInSats,
