+++ suite-common/wallet-core/src/blockchain/blockchainThunks.ts
@@ -3,13 +3,11 @@ import { notificationsActions } from '@suite-common/toast-notifications';
 import {
     NetworkSymbol,
     externalBackendTypeNetworks,
-    getNetwork,
     getNetworkOptional,
     isNetworkSymbol,
     isTrezorInfraBasedNetwork,
-    networksCollection,
 } from '@suite-common/wallet-config';
-import type { Account, CustomBackend, NetworksFees } from '@suite-common/wallet-types';
+import type { Account, CustomBackend } from '@suite-common/wallet-types';
 import {
     findAccountDevice,
     findAccountsByDescriptor,
@@ -20,7 +18,6 @@ import {
     getAreSatoshisUsed,
     getBackendFromSettings,
     getCustomBackends,
-    isEip1559,
     isTrezorConnectBackendType,
     shouldSubscribeBlocks,
     shouldUseIdentities,
@@ -29,7 +26,6 @@ import TrezorConnect, {
     BlockchainBlock,
     BlockchainError,
     BlockchainNotification,
-    FeeLevel,
 } from '@trezor/connect';
 import type { TimerId } from '@trezor/type-utils';
 import { arrayDistinct, arrayToDictionary } from '@trezor/utils';
@@ -38,7 +34,7 @@ import { BLOCKCHAIN_MODULE_PREFIX, blockchainActions } from './blockchainActions
 import { selectBlockchainState, selectNetworkBlockchainInfo } from './blockchainReducer';
 import { selectAccounts } from '../accounts/accountsReducer';
 import { fetchAndUpdateAccountThunk } from '../accounts/accountsThunks';
-import { selectNetworkFeeInfo } from '../fees/feesReducer';
+import { preloadFeeInfoThunk, updateFeeInfoThunk } from '../fees/feesThunks';
 
 export const DEFAULT_ACCOUNT_SYNC_INTERVAL = 60 * 1000; // 1 minute
 
@@ -55,155 +51,6 @@ const CUSTOM_ACCOUNT_SYNC_INTERVALS: Partial<Record<NetworkSymbol, number>> = {
 
 const getAccountSyncInterval = (symbol: NetworkSymbol) =>
     CUSTOM_ACCOUNT_SYNC_INTERVALS[symbol] || DEFAULT_ACCOUNT_SYNC_INTERVAL;
-
-// Conditionally subscribe to blockchain backend
-// called after TrezorConnect.init successfully emits TRANSPORT.START event
-// checks if there are discovery processes loaded from LocalStorage
-// if so starts subscription to proper networks
-
-// sort FeeLevels in reversed order (Low > High)
-// TODO: consider to use same order in @trezor/connect to avoid double sorting
-const order: FeeLevel['label'][] = ['low', 'economy', 'normal', 'high'];
-const sortLevels = (levels: FeeLevel[]) =>
-    levels.sort((levelA, levelB) => order.indexOf(levelA.label) - order.indexOf(levelB.label));
-
-// shouldn't this be in fee thunks instead?
-export const preloadFeeInfoThunk = createThunk(
-    `${BLOCKCHAIN_MODULE_PREFIX}/preloadFeeInfoThunk`,
-    async (_, { dispatch }) => {
-        // Fetch default fee levels
-        const networks = networksCollection.filter(n => !n.isHidden);
-        const promises = networks.map(network =>
-            TrezorConnect.blockchainEstimateFee({
-                coin: network.symbol,
-                request: {
-                    feeLevels: 'preloaded',
-                },
-            }),
-        );
-        const levels = await Promise.all(promises);
-
-        const partial: Partial<NetworksFees> = {};
-        networks.forEach((network, index) => {
-            const result = levels[index];
-
-            if (result.success) {
-                const { payload } = result;
-                partial[network.symbol] = {
-                    blockHeight: 0,
-                    ...payload,
-                    levels: sortLevels(
-                        payload.levels
-                            // hack to hide "low" fee option
-                            // (we do not want to change the connect API as it is a potentially breaking change)
-                            .filter(level => level.label !== 'low'),
-                    ).map(level => ({
-                        ...level,
-                        label: level.label || 'normal',
-                    })),
-                };
-            }
-        });
-
-        dispatch(blockchainActions.updateFee(partial));
-    },
-);
-
-// shouldn't this be in fee thunks instead?
-export const updateFeeInfoThunk = createThunk(
-    `${BLOCKCHAIN_MODULE_PREFIX}/updateFeeInfoThunk`,
-    async (
-        { networkSymbol, forceUpdate }: { networkSymbol: string; forceUpdate?: boolean },
-        { dispatch, getState, extra },
-    ) => {
-        const network = getNetworkOptional(networkSymbol.toLowerCase());
-        if (!network) return;
-        const blockchainInfo = selectNetworkBlockchainInfo(getState(), network.symbol);
-        const feeInfo = selectNetworkFeeInfo(getState(), network.symbol);
-        const device = extra.selectors.selectDevice(getState());
-        // TODO: remove when EIP-1559 complete
-        const isDebugModeActive = extra.selectors.selectDebugSettings(getState()).showDebugMenu;
-
-        if (
-            feeInfo &&
-            feeInfo.blockHeight > 0 &&
-            blockchainInfo.blockHeight - feeInfo.blockHeight < 10 &&
-            !forceUpdate
-        ) {
-            return;
-        }
-
-        let newFeeInfo;
-
-        if (network.networkType === 'ethereum') {
-            const result = await TrezorConnect.blockchainEstimateFee({
-                coin: network.symbol,
-                request: {
-                    blocks: [2],
-                    feeLevels: 'smart',
-                    specific: {
-                        from: '0x0000000000000000000000000000000000000000',
-                        to: '0x0000000000000000000000000000000000000000',
-                    },
-                },
-            });
-            if (result.success) {
-                const feeLevelBase = result.payload.levels[0];
-                const isEip1559ActivatedAndAvailable =
-                    getNetwork(network.symbol).features.includes('eip1559') &&
-                    isEip1559(feeLevelBase) &&
-                    isDebugModeActive &&
-                    !device?.unavailableCapabilities?.['eip1559'];
-
-                if (isEip1559ActivatedAndAvailable) {
-                    newFeeInfo = result.payload;
-                } else {
-                    newFeeInfo = {
-                        ...result.payload,
-                        levels: [
-                            {
-                                ...feeLevelBase,
-                                baseFeePerGas: undefined,
-                                maxFeePerGas: undefined,
-                                maxPriorityFeePerGas: undefined,
-                                maxWaitTimeEstimate: undefined,
-                                label: 'normal' as const,
-                            },
-                        ],
-                    };
-                }
-            }
-        } else {
-            const result = await TrezorConnect.blockchainEstimateFee({
-                coin: network.symbol,
-                request: {
-                    feeLevels: 'smart',
-                },
-            });
-            if (result.success) {
-                newFeeInfo = {
-                    ...result.payload,
-                    levels: sortLevels(
-                        result.payload.levels
-                            // hack to hide "low" fee option
-                            // (we do not want to change the connect API as it is a potentially breaking change)
-                            .filter(level => level.label !== 'low'),
-                    ),
-                };
-            }
-        }
-
-        if (newFeeInfo) {
-            const partial: Partial<NetworksFees> = {};
-            partial[network.symbol] = {
-                blockHeight: blockchainInfo.blockHeight,
-                ...newFeeInfo,
-            };
-
-            dispatch(blockchainActions.updateFee(partial));
-        }
-    },
-);
 
 // call TrezorConnect.unsubscribe, it doesn't cost anything and should emit BLOCKCHAIN.CONNECT or BLOCKCHAIN.ERROR event
 export const reconnectBlockchainThunk = createThunk(
