+++ suite-common/wallet-core/src/index.ts
@@ -16,7 +16,9 @@ export * from './discovery/discoveryActions';
 export * from './discovery/discoveryReducer';
 export * from './discovery/discoveryThunks';
 export * from './discovery/discoveryRunningStateLocks';
+export * from './fees/feesActions';
 export * from './fees/feesReducer';
+export * from './fees/feesThunks';
 export * from './fiat-rates/fiatRatesMiddleware';
 export * from './fiat-rates/fiatRatesReducer';
 export * from './fiat-rates/fiatRatesSelectors';
